requirements.txt
gin-config
matplotlib
numpy
scipy
absl-py
tensorflow
tensorboard
